* renaming files to their class contents
